+++ src/main/java/fr/alasdiablo/diolib/data/DioLootTableProvider.java
@@ -25,12 +25,16 @@ public class DioLootTableProvider extends LootTableProvider {
     private final DataGenerator dataGenerator;
     private static final Gson GSON = (new GsonBuilder()).setPrettyPrinting().disableHtmlEscaping().create();
     protected final List<Pair<Supplier<Consumer<BiConsumer<ResourceLocation, LootTable.Builder>>>, LootParameterSet>> lootTableList;
+    private final LootParameterSet lootParameterSet;
+    private final String name;
 
 
-    public DioLootTableProvider(DataGenerator dataGeneratorIn, List<Pair<Supplier<Consumer<BiConsumer<ResourceLocation, LootTable.Builder>>>, LootParameterSet>> lootTableListIn) {
+    public DioLootTableProvider(DataGenerator dataGeneratorIn, List<Pair<Supplier<Consumer<BiConsumer<ResourceLocation, LootTable.Builder>>>, LootParameterSet>> lootTableListIn, LootParameterSet lootParameterSetIn, String nameIn) {
         super(dataGeneratorIn);
         this.dataGenerator = dataGeneratorIn;
         this.lootTableList = lootTableListIn;
+        this.lootParameterSet = lootParameterSetIn;
+        this.name = nameIn;
     }
 
     @Override
@@ -44,8 +48,8 @@ public class DioLootTableProvider extends LootTableProvider {
             }
         }));
 
-        ValidationTracker validationtracker = new ValidationTracker(LootParameterSets.GENERIC, (resourceLocation) -> null, map::get);
-        validate(map, validationtracker);
+        ValidationTracker validationtracker = new ValidationTracker(this.lootParameterSet, (resourceLocation) -> null, map::get);
+        this.validate(map, validationtracker);
         Multimap<String, String> multimap = validationtracker.getProblems();
 
         if (!multimap.isEmpty()) {
@@ -66,11 +70,21 @@ public class DioLootTableProvider extends LootTableProvider {
     }
 
     @Override
+    protected void validate(Map<ResourceLocation, LootTable> map, ValidationTracker validationtracker) {
+        map.forEach((resourceLocation, lootTable) -> LootTableManager.validateLootTable(validationtracker, resourceLocation, lootTable));
+    }
+
+    @Override
     protected List<Pair<Supplier<Consumer<BiConsumer<ResourceLocation, LootTable.Builder>>>, LootParameterSet>> getTables() {
         return this.lootTableList;
     }
 
     private static Path getPath(Path pathIn, ResourceLocation id) {
         return pathIn.resolve("data/" + id.getNamespace() + "/loot_tables/" + id.getPath() + ".json");
+    }
+
+    @Override
+    public String getName() {
+        return this.name + " - LootTables";
     }
 }