+++ changelog.md
@@ -1,7 +1,7 @@
 ### 1.0.2
 + change mod id
 + change package name
-+ name containing only 6 letters 3 for the mod base name `Diabolo` and 3 from `Lib`
++ name containing only 6 letters, 3 from the mod name `Diabolo` and 3 from `Lib`
 
 ### 1.0.1
 + Add loot table provider